Move image size validation functions into separate file
